change empty lists to none
suggestion fo 8__ on reddit who notes:

"You should avoid having an empty list as one of the default arguments in a function, because it could lead to unexpected behaviour. Instead, try making the default None and then if it's None, then assign it an empty list.

The reason is because next time you run the function, the same list will be used."